foes that regenerate in dark were regenerating in daylight
The regeneration message says that the monster uses “the darkness” to
heal itself, so why does the game allow it to happen in full sunlight?

The darkness feature was balanced so that a floodlight prevents these
monsters from healing in a 20–tile radius. Floodlights create a light
level of 8000, which drops off as you get further away. 20 tiles away
the level has dropped to about 200. Unfortunately, full daylight at
noon on a clear day has an ambient light level of only around 127.5,
so while the monster may work fine in underground areas with the
occasional floodlight, it is not so great outdoors or in buildings
with windows.

This fix rebalances it so that the regeneration rate drops to zero at
a light level of 30, the same level where the (unrelated) PHOTOPHOBIA
effect kicks in. This lets the player use daylight from a window to
effectively neutralize this healing ability. Unfortunately other light
sources, like flashlights and floodlights, are even more effective
than daylight. Rebalancing those against the light from the sun,
however, seems like a rather more ambitious fix.

I also added a few additional messages for different light levels.

Fixes #67027